+++ src/drive.h
@@ -2,8 +2,8 @@
 #define DRIVE_H
 #include "variables.h"
 #include <structs.h>
-void drive(float leftSpeed, float rightSpeed);
-float calculateSpeedDelta(struct motor servo);
+void drive(double leftSpeed, double rightSpeed);
+double calculateSpeedDelta(struct motor servo);
 void updateSpeeds();
-void drive2(int absoluteX, int absoluteY, bool backward = false);
+void drive2(double absoluteX, double absoluteY, bool backward = false);
 #endif
